rename & refactor : MindMapView->MindMapCanvas 파일 명 변경 & ControlSection, MindMapCanvas, Minutes MindMapView로 묶기
