+++ client/src/components/MindMapCanvas/DeleteConfirmModal.tsx
