+++ client/src/components/MindMapCanvas/NoNodeInform.tsx
